Consolidate duplicate iOS/macOS plugin code into shared Swift source
Move the nearly identical AudioDecoderPlugin.swift into darwin/Classes/
with #if os(iOS)/#if os(macOS) guards for the two platform differences
(imports and messenger accessor). Replace the per-platform copies with
symlinks so both podspecs continue to resolve Classes/**/* as before.
